add support for 64-bit scsynth and plugins on OS X
git-svn-id: https://supercollider.svn.sourceforge.net/svnroot/supercollider/trunk@9377 a380766d-ff14-0410-b294-a243070f3f08

+++ Headers/server/SC_ComPort.h
@@ -58,7 +58,7 @@ class SC_ComPort : public SC_CmdPort
 {
 protected:
 	int mPortNum;
-	int mSocket;
+	int64 mSocket;
 	struct sockaddr_in mBindSockAddr;
 
 #ifdef USE_RENDEZVOUS
